rename metadata
