+++ transport/grpc/server.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/go-kratos/kratos/v2/api/kratos/api"
-	"github.com/go-kratos/kratos/v2/internal/desc"
+	"github.com/go-kratos/kratos/v2/internal/api/metadata"
 	"github.com/go-kratos/kratos/v2/internal/host"
 	"github.com/go-kratos/kratos/v2/log"
 	"github.com/go-kratos/kratos/v2/middleware"
@@ -79,7 +79,7 @@ type Server struct {
 	middleware middleware.Middleware
 	grpcOpts   []grpc.ServerOption
 	health     *health.Server
-	metaServer *desc.Server
+	metaServer *metadata.Server
 }
 
 // NewServer creates a gRPC server by options.
@@ -106,11 +106,11 @@ func NewServer(opts ...ServerOption) *Server {
 		grpcOpts = append(grpcOpts, srv.grpcOpts...)
 	}
 	srv.Server = grpc.NewServer(grpcOpts...)
-	srv.metaServer = desc.NewServer(srv.Server)
+	srv.metaServer = metadata.NewServer(srv.Server)
 	// grpc health register
 	healthpb.RegisterHealthServer(srv.Server, srv.health)
 	// api metadata register
-	api.RegisterDescriptionServer(srv.Server, srv.metaServer)
+	api.RegisterMetadataServer(srv.Server, srv.metaServer)
 	// reflection register
 	reflection.Register(srv.Server)
 	return srv
